requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
celery-progress==0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.3
distro==1.8.0
Django==4.2.8
django-celery-results==2.5.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
inflection==0.5.1
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.4
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.74
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.6.0
packaging==23.2
pgvector==0.2.4
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
yarl==1.9.4
